Business logic updated, Refactoring, multiple classes added

+++ Get-A-Taxi/app/src/main/java/com/getataxi/client/comm/models/ClientOrderDM.java
@@ -14,7 +14,7 @@ public class ClientOrderDM {
     public int orderId;
 
     @SerializedName("startLat")
-    public int orderLatitude;
+    public double orderLatitude;
 
     @SerializedName("startLng")
     public double orderLongitude;
